minor renaming + documentation update
